+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- File validation error does not prevent from submission. [#318](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/318)
+
 ## [2.7.0](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/milestone/12) 2022-09-27
 
 ### Added